All tests for services is done.

+++ DbAssignment/Services/OrderItemService.cs
@@ -14,9 +14,15 @@ public class OrderItemService(OrderItemRepository orderItemRepository, OrderServ
 
     public OrderItemEntity CreateOrderItem(int quantity, int orderId, int productId)
     {
-       
 
-        
+        var order = _orderService.GetOrderById(orderId);
+        var product = _productService.GetProductById(productId);
+
+        if (order == null || product == null)
+        {
+            return null;
+        }
+
         var orderItemEntity = new OrderItemEntity
         {
             Quantity = quantity,
@@ -27,7 +33,7 @@ public class OrderItemService(OrderItemRepository orderItemRepository, OrderServ
         orderItemEntity = _orderItemRepository.Create(orderItemEntity);
 
         return orderItemEntity;
-      
+
     }
 
     public OrderItemEntity GetOrderItemById(int id)
@@ -50,14 +56,31 @@ public class OrderItemService(OrderItemRepository orderItemRepository, OrderServ
 
     public OrderItemEntity UpdateOrderItem(OrderItemEntity orderItemEntity)
     {
-        var updatedOrderItem = _orderItemRepository.Update(x => x.Id == orderItemEntity.Id, orderItemEntity);
-        return updatedOrderItem;
+        var existingOrderItem = _orderItemRepository.Update(x => x.Id == orderItemEntity.Id, orderItemEntity);
+        if (existingOrderItem == null)
+        {
+            return null;
+        }
+
+        existingOrderItem.Quantity = orderItemEntity.Quantity;
+        existingOrderItem.OrderId = orderItemEntity.OrderId;
+        existingOrderItem.ProductId = orderItemEntity.ProductId;
+
+        _orderItemRepository.Update(x => x.Id == existingOrderItem.Id, existingOrderItem);
+        return existingOrderItem;
     }
 
 
-    public void DeleteOrderItem(int Id)
+    public bool DeleteOrderItem(int Id)
     {
+        var existingOrderItem = _orderItemRepository.Get(x => x.Id == Id);
+        if (existingOrderItem == null)
+        {
+            return false;
+        }
+
         _orderItemRepository.Delete(x => x.Id == Id);
+        return true;
 
     }
 }
